Moved key generator into more sensibly spot
